Remove flipmaps class
It is now in the view options class.
