+++ trview.app/UI/ViewOptions.h
@@ -2,6 +2,9 @@
 
 #include <cstdint>
 #include <trview.common/Event.h>
+#include <set>
+
+#include <trview.common/TokenStore.h>
 
 namespace trview
 {
@@ -10,6 +13,9 @@ namespace trview
         class Control;
         class Checkbox;
         class NumericUpDown;
+        class Button;
+        class Window;
+        class StackPanel;
     }
 
     struct ITextureStorage;
@@ -18,6 +24,9 @@ namespace trview
     {
     public:
         explicit ViewOptions(ui::Control& parent, const ITextureStorage& texture_storage);
+
+        /// Event raised when an alternate group is toggled. The group number and the new state are passed as parameters.
+        Event<uint16_t, bool> on_alternate_group;
 
         /// Event raised when the user has changed the depth of neighbour to display. The newly selected depth is passed
         /// when the event is raised.
@@ -53,6 +62,15 @@ namespace trview
         /// @remarks This event is not raised by the set_show_water function.
         Event<bool> on_show_water;
 
+        /// Set whether an alternate group is enabled. This will not raise the on_alternate_group event.
+        /// @param value The group to change.
+        /// @param enabled Whether the group is enabled.
+        void set_alternate_group(uint16_t value, bool enabled);
+
+        /// Set the alternate groups that are in the level.
+        /// @param groups The groups in the level.
+        void set_alternate_groups(const std::set<uint16_t>& groups);
+
         /// Set the value of the depth control. This will not raise the on_depth_changed event.
         /// @param value The neighbour depth to use.
         void set_depth(int32_t value);
@@ -60,6 +78,14 @@ namespace trview
         /// Set whether neighbours are enabled. This will not raise the on_enabled_changed event.
         /// @param value Whether neighbours are enabled.
         void set_depth_enabled(bool value);
+
+        /// Set the current flip mode. This will not raise the on_flip event but will update the user interface appropriately.
+        /// @param flip The new flip mode.
+        void set_flip(bool flip);
+
+        /// Set whether the flip control is enabled or disabled.
+        /// @param enabled Whether the control is enabled.
+        void set_flip_enabled(bool enabled);
 
         /// Set the current highlight mode. This will not raise the on_highlight event but will update the user interface appropriately.
         /// @param highlight Whether the highlight mode is enabled or disabled.
@@ -77,6 +103,10 @@ namespace trview
         /// @param show Whether water is visible.
         void set_show_water(bool show);
 
+        /// Set whether to use alternate groups method of flipmaps.
+        /// @param value Whether to use alternate groups or a single toggle.
+        void set_use_alternate_groups(bool value);
+
         /// Get the current value of the show hidden geometry checkbox.
         /// @returns The current value of the checkbox.
         bool show_hidden_geometry() const;
@@ -89,12 +119,18 @@ namespace trview
         /// @returns The current value of the checkbox.
         bool show_water() const;
     private:
-        ui::Checkbox*      _highlight;
-        ui::Checkbox*      _flip;
-        ui::Checkbox*      _triggers;
-        ui::Checkbox*      _hidden_geometry;
-        ui::Checkbox*      _water;
-        ui::Checkbox*      _enabled;
+        TokenStore _token_store;
+        ui::Checkbox* _highlight;
+        ui::Checkbox* _flip;
+        ui::Checkbox* _triggers;
+        ui::Checkbox* _hidden_geometry;
+        ui::Checkbox* _water;
+        ui::Checkbox* _enabled;
         ui::NumericUpDown* _depth;
+        ui::Window* _tr1_3_panel;
+        ui::Window* _tr4_5_panel;
+        ui::StackPanel* _alternate_groups;
+        std::unordered_map<uint16_t, bool> _alternate_group_values;
+        std::unordered_map<uint16_t, ui::Button*> _alternate_group_buttons;
     };
 }
