+++ trview.app/UI/ViewerUI.h
@@ -14,7 +14,6 @@
 #include <trview.app/UI/RoomNavigator.h>
 #include <trview.app/UI/GoTo.h>
 #include <trview.app/UI/SettingsWindow.h>
-#include <trview.app/UI/Flipmaps.h>
 #include <trview.app/UI/CameraControls.h>
 #include <trview.app/UI/CameraPosition.h>
 #include <trview.app/Geometry/PickInfo.h>
@@ -278,7 +277,6 @@ namespace trview
         std::unique_ptr<Toolbar> _toolbar;
         std::unique_ptr<LevelInfo> _level_info;
         std::unique_ptr<SettingsWindow> _settings_window;
-        std::unique_ptr<Flipmaps> _flipmaps;
         std::unique_ptr<CameraControls> _camera_controls;
         std::unique_ptr<CameraPosition> _camera_position;
         std::unique_ptr<ui::render::MapRenderer> _map_renderer;
